third_party/WebKit/LayoutTests/fast/css/children-affected-by-style-sharing-expected.txt

childrenOrSiblingsAffectedBy* disables style sharing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(span).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE


You should see a green rectangle below.

FAIL
